backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.13.0
backoff==2.2.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
distro==1.9.0
fastapi==0.115.0
greenlet==3.4.0
groq==0.11.0
grpcio==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
jiter==0.14.0
mem0ai==1.0.11
numpy==2.4.4
openai==2.31.0
portalocker==3.2.0
posthog==7.11.1
protobuf==6.33.6
pydantic==2.13.0
pydantic_core==2.46.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
qdrant-client==1.17.1
requests==2.33.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==0.38.6
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.30.6
